Fix velocity handling in envelopes

+++ src/generic_output.h
@@ -111,7 +111,7 @@ struct GenOut_t : GenOut_base {
 		struct {
 			EnvStage_t envelope_stage;
 			uint32_t env_value;
-			uint32_t env_velocity;
+			uint16_t env_velocity;
 			uint16_t env_sustain;
 		};
 	};
@@ -284,7 +284,7 @@ class key_handler_c {
 class env_handler_c {
 	public:
 	void handle_cvm(umpCVM* msg);
-	uint32_t get(EnvStage_t stage, uint32_t vel);
+	uint32_t get(EnvStage_t stage, uint16_t vel);
 	void set_go(GenOut_t* handler, EnvStage_t stage);
 	bool enabled(EnvStage_t stage);
 	Env_t env;
